requirements.txt
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
bandit==1.6.2
bleach==3.1.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
codecov==2.0.22
colorama==0.4.3
coverage==5.0.4
distlib==0.3.0
docutils==0.16
dparse==0.5.0
filelock==3.0.12
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
isort==4.3.21
keyring==21.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pbr==5.4.4
Pillow==7.0.0
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
Pygments==2.6.1
pyinstrument==3.1.3
pyinstrument-cext==0.2.2
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
safety==1.8.7
six==1.14.0
smmap==3.0.1
snowballstemmer==2.0.0
stevedore==1.32.0
toml==0.10.0
tqdm==4.43.0
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.13
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
